Added sign-out feature
