+++ app/views/layout.erb
@@ -5,6 +5,7 @@
   <link rel="stylesheet" type="text/css" href="/css/application.css">
 </head>
 <body>
+
   <% if current_user %>
     Welcome, <%= current_user.name %>
   <% end %>
@@ -12,10 +13,23 @@
     Sorry, there were the following problems with the form.
     <ul id="errors">
       <% flash[:errors].each do |error| %>
-        <li><%= error %></li>
+      <li><%= error %></li>
       <% end %>
     </ul>
 <% end %>
+
+    <% if current_user %>
+  <form method="post" action="/sessions">
+    <input type="hidden" name="_method" value="delete">
+    <input type="submit" value="Sign out">
+  </form>
+  <% if flash[:notice] %>
+    <div id="notice">
+      <%= flash[:notice] %>
+    </div>
+  <% end %>
+<% end %>
+
   <%= yield %>
 </body>
 </html>
